+++ unity/Runtime/Bindings/MjBindings.cs
@@ -573,6 +573,18 @@ public enum mjtOrientation : int{
   mjORIENTATION_ZAXIS = 3,
   mjORIENTATION_EULER = 4,
 }
+public enum mjtCTimer : int{
+  mjCTIMER_TOTAL = 0,
+  mjCTIMER_ASSETS = 1,
+  mjCTIMER_TEXTURE = 2,
+  mjCTIMER_MESH_LOAD = 3,
+  mjCTIMER_MESH_HULL = 4,
+  mjCTIMER_MESH_POLYGON = 5,
+  mjCTIMER_MESH_INERTIA = 6,
+  mjCTIMER_MESH_BVH = 7,
+  mjCTIMER_MESH_OCTREE = 8,
+  mjNCTIMER = 9,
+}
 public enum mjtCatBit : int{
   mjCAT_STATIC = 1,
   mjCAT_DYNAMIC = 2,